+++ Dockerfile
@@ -12,7 +12,7 @@ RUN export PATH=$PATH:/usr/bin/python3.11
 # Install curl (and vim if you need it)
 # RUN apt-get update && apt-get install -y curl vim
 RUN pip install requests
-RUN pip install neo4j 
+RUN pip install neo4j
 RUN pip install owlready2
 RUN pip install rdflib
 
@@ -20,10 +20,10 @@ RUN pip install rdflib
 COPY . /app
 
 # Set permissions for the JAR and data directory
-RUN chmod 755 /app/rml_mapper/mapper.jar
+RUN chmod 755 /app/mapping/mapper.jar
 
 # Set environment variables
-ENV UCO_ONTO_PATH=/app/uco2.ttl
+ENV UCO_ONTO_PATH=/app/data/UCKG_Snapshots/uco2.ttl
 ENV ROOT_FOLDER=/app
 
 # Set the Python script as the entry point
