--- att&ck_research/READ_ME.txt
@@ -1,1 +1,0 @@
-This folder contains research documents for integrating Mitre ATT&CK into the UCKG, specifically in relation to CVE.
