+++ mapping/cve/cve_rml.ttl
@@ -1,0 +1,108 @@
+@prefix rml: <http://semweb.mmlab.be/ns/rml#> .
+@prefix rr: <http://www.w3.org/ns/r2rml#> .
+@prefix ql: <http://semweb.mmlab.be/ns/ql#> .
+@prefix rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#> .
+@prefix : <http://example.org/rules/> .
+@prefix schema: <http://schema.org/> .
+@prefix dbo: <http://dbpedia.org/ontology/> .
+@prefix cve: <http://purl.org/cyber/cve#> .
+@prefix owl: <http://www.w3.org/2002/07/owl#> .
+@prefix stx: <http://purl.org/cyber/stx#> .
+@prefix uco: <http://purl.org/cyber/uco#> .
+@prefix xml: <http://www.w3.org/XML/1998/namespace> .
+@prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
+@prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
+@prefix capec: <http://purl.org/cyber/capec#> .
+@base <http://purl.org/cyber/uco#> .
+
+# How to read the CVEs from a json file and iterate
+# Defining a source json file as a ClassAMap (Arbitrary naming scheme)
+# Eventually declaring the ClassAMap as a TriplesMap Object
+uco:ClassAMap a rr:TriplesMap;
+  rml:logicalSource [
+    rml:source "./mapping/cve/cves.json";
+    rml:referenceFormulation ql:JSONPath;
+    rml:iterator "cves[*]"
+  ];
+
+#Vulnerability Class mapping
+rr:subjectMap [
+  rr:template "http://purl.org/cyber/uco#VULN-{cve.id}";
+  rr:class uco:ucoVulnerability
+].
+
+#CVE Class mapping
+uco:ClassBMap a rr:TriplesMap;
+  rml:logicalSource [
+    rml:source "./mapping/cve/cves.json";
+    rml:referenceFormulation ql:JSONPath;
+    rml:iterator "cves[*]"
+  ];
+rr:subjectMap [
+  rr:template "http://purl.org/cyber/uco#{cve.id}";
+  rr:class uco:ucoCVE
+].
+
+#CWE Weakness Class mapping
+uco:ClassCMap a rr:TriplesMap;
+  rml:logicalSource [
+    rml:source "./mapping/cve/cves.json";
+    rml:referenceFormulation ql:JSONPath;
+    rml:iterator "$.cves[*].cve.cwes[*].cwe"
+  ];
+rr:subjectMap [
+  rr:template "http://purl.org/cyber/uco#ExploitTarget-{id}";
+  rr:class uco:ucoExploitTarget
+].
+
+# Vulnerability hasCVE_ID predicate
+# This is also how you join two classes
+uco:ClassAMap rr:predicateObjectMap [
+    rr:predicate uco:ucohasVulnerability;  # The predicate linking ClassA to ClassB
+    rr:objectMap [
+        rr:parentTriplesMap uco:ClassBMap;  # Referencing the ClassB map
+        rr:joinCondition [
+          rr:child "cve.id";
+          rr:parent "cve.id"
+        ]
+    ]
+].
+
+uco:ClassCMap rr:predicateObjectMap [
+    rr:predicate uco:ucohasVulnerability;  # The predicate linking ClassA to ClassB
+    rr:objectMap [
+        rr:parentTriplesMap uco:ClassAMap;  # Referencing the ClassB map
+        rr:joinCondition [
+          rr:child "cve.id";
+          rr:parent "id"
+        ]
+    ]
+].
+
+
+#Vulnerability ucoPublishedDate predicate
+uco:ClassAMap rr:predicateObjectMap [
+        rr:predicate uco:ucoPublishedDate;  # Your DatatypeProperty for dateTime
+        rr:objectMap [
+            rml:reference "cve.published";  # Field in your data source
+            rr:datatype xsd:dateTime  # Specifying that the datatype is dateTime
+        ]
+].
+
+#Vulnerability ucoLastModifiedDate predicate
+uco:ClassAMap rr:predicateObjectMap [
+        rr:predicate uco:ucoLastModifiedDate;  # Your DatatypeProperty for dateTime
+        rr:objectMap [
+            rml:reference "cve.lastModified";  # Field in your data source
+            rr:datatype xsd:dateTime  # Specifying that the datatype is dateTime
+        ]
+].
+
+#Vulnerability ucosummary predicate
+uco:ClassAMap rr:predicateObjectMap [
+        rr:predicate uco:ucosummary;  # Your DatatypeProperty for dateTime
+        rr:objectMap [
+            rml:reference "cve.descriptions[*].value";  # Field in your data source
+            rr:datatype xsd:string  # Specifying that the datatype is dateTime
+        ]
+].
