+++ save/att&ck_research/READ_ME.txt
@@ -1,0 +1,1 @@
+This folder contains research documents for integrating Mitre ATT&CK into the UCKG, specifically in relation to CVE.
